Refactor transaction-details view to Typescript
